+++ spring-cloud-task-core/src/main/java/org/springframework/cloud/task/configuration/SimpleTaskConfiguration.java
@@ -28,6 +28,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.ApplicationArguments;
 import org.springframework.cloud.task.listener.TaskLifecycleListener;
 import org.springframework.cloud.task.listener.annotation.TaskListenerExecutor;
+import org.springframework.cloud.task.listener.annotation.TaskListenerExecutorFactory;
 import org.springframework.cloud.task.repository.TaskExplorer;
 import org.springframework.cloud.task.repository.TaskNameResolver;
 import org.springframework.cloud.task.repository.TaskRepository;
@@ -74,9 +75,11 @@ public class SimpleTaskConfiguration {
 	}
 
 	@Bean
-	public TaskListenerExecutor taskListenerExecutor(ConfigurableApplicationContext context)
+	public TaskListenerExecutor taskListenerExecutor(ConfigurableApplicationContext context) throws Exception
 	{
-		return new TaskListenerExecutor(context);
+		TaskListenerExecutorFactory taskListenerExecutorFactory =
+				new TaskListenerExecutorFactory(context);
+		return taskListenerExecutorFactory.getObject();
 	}
 	
 	@Bean
